+++ cmake/modules/Element.cmake
@@ -36,6 +36,8 @@ target_sources(kwik_element
             modules/element/g3d.cppm
             modules/element/scroll_view.cppm
             modules/element/tree_menu.cppm
+            modules/element/lazy_list_source.cppm
+            modules/element/lazy_list.cppm
 
         BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/modules/layout
         FILES
@@ -80,6 +82,7 @@ target_sources(kwik_element
         src/element/g3d.cpp
         src/element/scroll_view.cpp
         src/element/tree_menu.cpp
+        src/element/lazy_list.cpp
 
         src/layout/flex_layout.cpp
         src/layout/grid_layout.cpp